refactor: rename tosshin gazebo plugins into dienen gazebo plugins
