+++ README.md
@@ -1,12 +1,12 @@
-# Tosshin Gazebo Plugins
+# Dienen Gazebo Plugins
 
-[![latest version](https://img.shields.io/github/v/release/threeal/tosshin_gazebo_plugins)](https://github.com/threeal/tosshin_gazebo_plugins/releases/)
-[![commits since latest version](https://img.shields.io/github/commits-since/threeal/tosshin_gazebo_plugins/latest)](https://github.com/threeal/tosshin_gazebo_plugins/commits/master)
-[![milestone](https://img.shields.io/github/milestones/progress/threeal/tosshin_gazebo_plugins/1?label=milestone)](https://github.com/threeal/tosshin_gazebo_plugins/milestone/1)
-[![license](https://img.shields.io/github/license/threeal/tosshin_gazebo_plugins)](./LICENSE)
-[![test status](https://img.shields.io/github/workflow/status/threeal/tosshin_gazebo_plugins/Build%20and%20Test?label=test)](https://github.com/threeal/tosshin_gazebo_plugins/actions)
+[![latest version](https://img.shields.io/github/v/release/threeal/dienen_gazebo_plugins)](https://github.com/threeal/dienen_gazebo_plugins/releases/)
+[![commits since latest version](https://img.shields.io/github/commits-since/threeal/dienen_gazebo_plugins/latest)](https://github.com/threeal/dienen_gazebo_plugins/commits/master)
+[![milestone](https://img.shields.io/github/milestones/progress/threeal/dienen_gazebo_plugins/1?label=milestone)](https://github.com/threeal/dienen_gazebo_plugins/milestone/1)
+[![license](https://img.shields.io/github/license/threeal/dienen_gazebo_plugins)](./LICENSE)
+[![test status](https://img.shields.io/github/workflow/status/threeal/dienen_gazebo_plugins/Build%20and%20Test?label=test)](https://github.com/threeal/dienen_gazebo_plugins/actions)
 
-This package provides a [Gazebo](http://gazebosim.org/) simulation's plugins implementation of [Tosshin](https://github.com/threeal/tosshin) mobile robot navigation project.
+This package provides a [Gazebo](http://gazebosim.org/) simulator's plugins for [Dienen](https://github.com/threeal/paper-simulasi-robot) assistive robot simulation.
 
 ## License
