+++ package.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>tosshin_gazebo_plugins</name>
+  <name>dienen_gazebo_plugins</name>
   <version>0.3.0</version>
-  <description>Gazebo plugins implementation of Tosshin mobile robot navigation project</description>
+  <description>Gazebo plugins for Dienen assistive robot simulation</description>
   <maintainer email="alfi.maulana.f@gmail.com">Alfi Maulana</maintainer>
   <license>MIT License</license>
   <buildtool_depend>ament_cmake</buildtool_depend>
